refactor(desktop): promote the Code page subcomponents into files
CodeDeliveryPage.tsx and CodeWorkspacePage.tsx bundled dozens of inline
subcomponents, so every delivery or workspace change landed in one of two
files. Each page now keeps only its composition root; the subcomponents,
list rows, filters, dialogs, badges, and helpers live under
src/code/delivery/ and src/code/workspace/ with the same names and no
behavior change. Existing exports stay reachable from the page modules.
